tests: refactor across multiple files
